Remove conductor plugin and move hook implementations
Eliminate the persistent CanaryConductor plugin now that resource-pool state is
owned by config.resource_manager. The conductor no longer needs to maintain
state across hooks.

Move the default canary_runtests implementation into runtest as a stateless
module-level hook. The default runner now constructs a ResourceQueue from the
resource pool provided by config.resource_manager and executes jobs through the
ResourceQueueExecutor.

Move the default canary_resource_pool_* hook implementations into
resource_pool/hooks alongside the existing resource_pool_fill/update hooks.
These implementations now delegate directly to config.resource_manager for
resource queries, descriptions, and accommodation checks.
